Rearrange experimental tests directories
